refactor(#2054): move rpc_codec from nexus.core to nexus.contracts
The proxy module imported decode_rpc_message from nexus.core.rpc_codec,
violating the NEXUS-LEGO-ARCHITECTURE brick rule that bricks must not
import from kernel internals. Since rpc_codec is a pure serialization
utility with zero kernel dependencies, move it to the tier-neutral
nexus.contracts package.

- Move src/nexus/core/rpc_codec.py → src/nexus/contracts/rpc_codec.py
- Update all 9 import sites (proxy, server, remote, services, tests)
- Re-export RPCEncoder, decode/encode_rpc_message from contracts __init__
- Fix logging overhead: module-level logger + isEnabledFor guards
- Delete old core/rpc_codec.py (no re-export shim)
