+++ src/nexus/proxy/transport.py
@@ -20,7 +20,7 @@ from tenacity import (
     wait_exponential_jitter,
 )
 
-from nexus.core.rpc_codec import decode_rpc_message
+from nexus.contracts.rpc_codec import decode_rpc_message
 from nexus.proxy.config import ProxyBrickConfig
 from nexus.proxy.errors import RemoteCallError
