+++ src/nexus/remote/async_client.py
@@ -53,6 +53,7 @@ from tenacity import (
     wait_exponential,
 )
 
+from nexus.contracts.rpc_codec import decode_rpc_message, encode_rpc_message
 from nexus.core.exceptions import (
     NexusFileNotFoundError,
     RemoteConnectionError,
@@ -60,7 +61,6 @@ from nexus.core.exceptions import (
     RemoteTimeoutError,
 )
 from nexus.core.filesystem import NexusFilesystem
-from nexus.core.rpc_codec import decode_rpc_message, encode_rpc_message
 from nexus.remote.base_client import BaseRemoteNexusFS
 from nexus.remote.rpc_proxy import RPCProxyBase
 from nexus.server.protocol import RPCRequest, RPCResponse
