+++ src/nexus/remote/client.py
@@ -51,6 +51,7 @@ from tenacity import (
     wait_exponential,
 )
 
+from nexus.contracts.rpc_codec import decode_rpc_message, encode_rpc_message
 from nexus.core.exceptions import (
     NexusFileNotFoundError,
     RemoteConnectionError,
@@ -58,7 +59,6 @@ from nexus.core.exceptions import (
     RemoteTimeoutError,
 )
 from nexus.core.filesystem import NexusFilesystem
-from nexus.core.rpc_codec import decode_rpc_message, encode_rpc_message
 from nexus.remote.base_client import BaseRemoteNexusFS
 from nexus.remote.rpc_proxy import RPCProxyBase
 from nexus.server.protocol import RPCRequest, RPCResponse
