+++ tests/unit/server/test_protocol.py
@@ -7,7 +7,7 @@ from datetime import datetime
 
 import pytest
 
-from nexus.core.rpc_codec import RPCEncoder, decode_rpc_message, encode_rpc_message
+from nexus.contracts.rpc_codec import RPCEncoder, decode_rpc_message, encode_rpc_message
 from nexus.server.protocol import (
     METHOD_PARAMS,
     RPCErrorCode,
